requirements.txt
django
djangorestframework
celery
redis
selenium
playwright
